refactor connection specific information to ConnectionHandler
